Refactor App into modular components with custom hooks
