+++ remaleh-protect-frontend/src/lib/api.js
@@ -1,2 +1,2 @@
 // Central API base for all fetch calls.
-export const API = import.meta.env.VITE_API_BASE || "http://localhost:10000";
+export const API = import.meta.env.VITE_API_BASE || import.meta.env.VITE_API_URL || "http://localhost:5001";
